+++ src/main/java/request/BuyMarketRequest.java
@@ -1,10 +1,11 @@
 package request;
 
+import java.math.BigDecimal;
+
 public class BuyMarketRequest {
     public String symbol;
 
     public String type;
 
-    public Double quantity;
-
+    public BigDecimal quantity;
 }
