+++ src/main/java/request/SellMarketRequest.java
@@ -1,9 +1,11 @@
 package request;
 
+import java.math.BigDecimal;
+
 public class SellMarketRequest {
     public String symbol;
 
     public String type;
 
-    public Double quantity;
+    public BigDecimal quantity;
 }
